Root, Admin and Data Store Classes.
This commit contains the sections I worked on, namely:

The Database Class
The Database Setup File
The Graph Class used in the test used by @Mthompson251 & @dhirensolanki
in the Test class.

The root application class `AppDelegate` that contains global functions
a compositions to other’s classes

The FlatUI ttk styling class.
and finally, the Admin interface and GUI handler class.